Refactored link processing logic in post analytics
ref https://linear.app/ghost/issue/PROD-2055/same-link-appears-multiple-times-in-link-list

- Links were grouped only on the Newsletter tab of post analytics, but not on the overview which resulted in similar links appearing multiple times. This PR factored out the link processing and grouping logic to link utilities so it's reusable at both places.
